refactor: rename configuration class name & package name to specific names
