+++ src/main/java/com/avalon/avalonchat/configuration/nurigo/NurigoProperties.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.configuration.message;
+package com.avalon.avalonchat.configuration.nurigo;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.ConstructorBinding;
@@ -8,13 +8,13 @@ import lombok.Getter;
 @Getter
 @ConstructorBinding
 @ConfigurationProperties(prefix = "cool-sms")
-public class CoolSmsProperties {
+public class NurigoProperties {
 	private final String apiKey;
 	private final String apiSecret;
 	private final String fromNumber;
 	private final String domain;
 
-	public CoolSmsProperties(String apiKey, String apiSecret, String fromNumber, String domain) {
+	public NurigoProperties(String apiKey, String apiSecret, String fromNumber, String domain) {
 		this.apiKey = apiKey;
 		this.apiSecret = apiSecret;
 		this.fromNumber = fromNumber.replaceAll("-", "").trim();
