+++ src/test/java/com/avalon/avalonchat/infra/message/configuration/NurigoPropertiesTest.java
@@ -6,12 +6,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.avalon.avalonchat.configuration.message.CoolSmsProperties;
+import com.avalon.avalonchat.configuration.nurigo.NurigoProperties;
 
 @SpringBootTest
-class CoolSmsPropertiesTest {
+class NurigoPropertiesTest {
 	@Autowired
-	private CoolSmsProperties properties;
+	private NurigoProperties properties;
 
 	@Test
 	void CoolSms_속성주입_성공() {
